LayoutTests/fast/forms/date/date-fixed-placeholder.html

<!DOCTYPE html>
<body>
<script src="../../js/resources/js-test-pre.js"></script>

<input type="date" id="date-without-placeholder-attribute">
<input type="date" id="date-with-placeholder-attribute" placeholder="Your birth day">

<script>
description('Check placeholder behavior of type=date with ENABLE_CALENDAR_PICKER');
if (window.internals) {
    var visiblePlaceholder = internals.visiblePlaceholder(document.getElementById('date-without-placeholder-attribute')).toLowerCase();
    // Assume English locale.
    shouldBeTrue('visiblePlaceholder.indexOf("year") >= 0');
    shouldBeTrue('visiblePlaceholder.indexOf("month") >= 0');
    shouldBeTrue('visiblePlaceholder.indexOf("day") >= 0');
    shouldBe('visiblePlaceholder', 'internals.visiblePlaceholder(document.getElementById("date-with-placeholder-attribute")).toLowerCase()');
} else
    debug('Both of the date fields should have default placeholders which contain "year", "month", and "day."');


</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
